增加配置文件

+++ mk_cert.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
+set -euo pipefail
+
+source .env
 
 mkdir -p ssl && cd "$_" || exit
 
-[ ! -f i.com.key ] && openssl genrsa -out i.com.key 4096
+if [ ! -f ca.crt ]; then
+    echo "请先生成CA证书"
+    exit 0
+fi
 
-[ ! -f i.com.csr ] && \
+[ ! -f "${HOST}".key ] && openssl genrsa -out "${HOST}".key 4096
+
+[ ! -f "${HOST}".csr ] && \
 openssl req -sha512 -new \
-    -subj "/C=CN/ST=Shanghai/L=Shanghai/O=iuxt/OU=iuxt/CN=i.com" \
-    -key i.com.key \
-    -out i.com.csr
+    -subj "/C=CN/ST=Shanghai/L=Shanghai/O=iuxt/OU=iuxt/CN=${HOST}" \
+    -key "${HOST}".key \
+    -out "${HOST}".csr
 
 
 cat > v3.ext <<-EOF
@@ -17,17 +25,17 @@ extendedKeyUsage = serverAuth, clientAuth
 subjectAltName=@SubjectAlternativeName
 
 [ SubjectAlternativeName ]
-DNS.1=i.com
-DNS.2=*.i.com
+DNS.1=${HOST}
+DNS.2=*.${HOST}
 DNS.3=localhost
-IP.1=127.0.0.1
-IP.2=10.0.0.30
-IP.3=10.0.0.3
+IP.1=${IP_1}
+IP.2=${IP_2}
+IP.3=${IP_3}
 EOF
 
-[ ! -f i.com.crt ] && \
+[ ! -f "${HOST}".crt ] && \
 openssl x509 -req -sha512 -days 3650 \
     -extfile v3.ext \
     -CA ca.crt -CAkey ca.key -CAcreateserial \
-    -in i.com.csr \
-    -out i.com.crt
+    -in "${HOST}".csr \
+    -out "${HOST}".crt
